Remove :id parameter in /users/:id/**/* and redefine them as /me/**/* .
