+++ app/controllers/me/tweet_import_progresses_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Me
+  class TweetImportProgressesController < ApplicationController
+    before_action :authenticate_user!
+
+    # GET /me/tweet_import_progress
+    def show
+      progress = TweetImportProgress.find_by!(user: current_user)
+      render json: progress
+    end
+  end
+end
