+++ app/controllers/users/statuses_controller.rb
@@ -2,8 +2,6 @@
 
 module Users
   class StatusesController < ApplicationController
-    before_action :authenticate_user!, except: %i|index|
-
     def index
       user = User.find(params[:user_id])
 
@@ -11,32 +9,6 @@ module Users
 
       statuses = CollectUserStatusesService.call!(**params_to_collect_by)
       render json: statuses
-    end
-
-    # POST /users/:id/statuses
-    def create
-      user = User.find(params[:user_id])
-      authorize_operation_for!(user)
-
-      if user.tweet_import_progress.present?
-        head :too_many_requests
-      else
-        MakeInitialTweetImportJob.perform_later(user_id: params[:user_id])
-        head :accepted
-      end
-    end
-
-    # PUT /users/:id/statuses
-    def update
-      user = User.find(params[:user_id])
-      authorize_operation_for!(user)
-
-      if user.tweet_import_lock.present?
-        head :too_many_requests
-      else
-        MakeAdditionalTweetImportJob.perform_later(user_id: params[:user_id])
-        head :accepted
-      end
     end
 
     private
