+++ config/routes.rb
@@ -28,21 +28,25 @@ Rails.application.routes.draw do
 
   resources :tweeted_dates, only: %i|index|
 
-  get :me, to: "users#me"
+  resources :users, only: %i|index| do
+    scope module: :users do
+      resources :statuses,      only: %i|index|
+      resources :tweeted_dates, only: %i|index|
+    end
+  end
   get "/users/:screen_name", to: "users#show", as: :user_page
 
-  resources :users, only: %i|index destroy| do
-    scope module: :users do
-      resources :statuses,              only: %i|index create|
-      put "statuses",                   to: "statuses#update"
-      resources :tweeted_dates,         only: %i|index|
-      resources :followees,             only: %i|create|
-      resource  :tweet_import_progress, only: %i|show|
+  resource :me, controller: "me/users", only: %i|show destroy|
+  namespace :me do
+    resources :statuses,              only: %i|index create|
+    put "statuses",                   to: "statuses#update"
+    resources :tweeted_dates,         only: %i|index|
+    resources :followees,             only: %i|create|
+    resource  :tweet_import_progress, only: %i|show|
 
-      namespace :followees do
-        resources :statuses,      only: %i|index|
-        resources :tweeted_dates, only: %i|index|
-      end
+    namespace :followees do
+      resources :statuses,      only: %i|index|
+      resources :tweeted_dates, only: %i|index|
     end
   end
